Relocate duplicated mock implementations to test fixtures
Closes gh-28659
